requirements.txt
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
tqdm>=4.65.0
